// src/lib/enums/product.enum.ts
export enum ProductSize {
    SMALL = "SMALL",
    NORMAL = "NORMAL",
    LARGE = "LARGE",
    EXTRA_LARGE = "EXTRA_LARGE",
    SET = "SET",
}

export enum ProductVolume {
    ZERO = "ZERO",
    HALF = "HALF",
    ONE = "ONE",
    ONE_POINT_FIVE = "ONE_POINT_FIVE",
    TWO = "TWO",
    THREE = "THREE",
    FIVE = "FIVE",
    TEN = "TEN",
}

export enum ProductStatus {
    PAUSE = "PAUSE",
    PROCESS = "PROCESS",
    DELETE = "DELETE",
}

export enum ProductCollection {
    EXTERIOR_CARE = "EXTERIOR_CARE",
    INTERIOR_CARE = "INTERIOR_CARE",
    WHEELS_TIRES = "WHEELS_TIRES",
    GLASS_MIRRORS = "GLASS_MIRRORS",
    TOOLS_ACCESSORIES = "TOOLS_ACCESSORIES",
    POLISH_CORRECTION = "POLISH_CORRECTION",
    OTHERS = "OTHERS",
}